+++ tests/dmock/asr9k-64/admin_show_install_inactive.txt
