+++ tests/dmock/asr9k-64/admin_show_inventory_chassis.txt
