+++ tests/dmock/asr9k-64/show_install_inactive.txt
