+++ tests/dmock/asr9k-64/show_users.txt
