+++ tests/dmock/asr9k-64/show_version.txt
