+++ tests/dmock/asr9k/admin_show_install_active.txt
