+++ tests/dmock/asr9k/admin_show_inventory_chassis.txt
