+++ tests/dmock/asr9k/show_install_active.txt
