+++ tests/dmock/asr9k/show_install_inactive.txt
