+++ tests/dmock/asr9k/show_version.txt
