+++ tests/dmock/asr9k/show_version_brief.txt
