+++ tests/dmock/n9k/show_inventory.txt
