+++ tests/dmock/n9k/show_inventory_chassis.txt
