+++ tests/dmock/n9k/show_users.txt
