+++ tests/dmock/n9k/show_version.txt
