+++ tests/dmock/n9k/show_version_brief.txt
