+++ tests/dmock/ncs1k/admin_show_install_active.txt
