+++ tests/dmock/ncs1k/admin_show_inventory_chassis.txt
