+++ tests/dmock/ncs1k/show_install_active.txt
