+++ tests/dmock/ncs1k/show_install_request_org.txt
