+++ tests/dmock/ncs1k/show_users.txt
