+++ tests/dmock/ncs1k/show_version.txt
