+++ tests/dmock/ncs1k/show_version_brief.txt
