+++ tests/dmock/ncs5500/admin_show_install_committed.txt
