+++ tests/dmock/ncs5500/admin_show_inventory_chassis.txt
