+++ tests/dmock/ncs5500/show_install_committed.txt
