+++ tests/dmock/ncs5500/show_version_brief.txt
