+++ tests/system/test_connection_asr9k.py
@@ -26,56 +26,47 @@
 # THE POSSIBILITY OF SUCH DAMAGE.
 # =============================================================================
 
-
-import os
-
 from unittest import TestCase
 
-from xrmock.xrmock import TelnetServer, XRHandler
+from tests.dmock.dmock import TelnetServer, ASR9KHandler
 from threading import Thread
 
 import condoor
+import os
 
 
-class NCS1KHandler(XRHandler):
-    platform = "NCS1K"
-    response_dict = {
-        'show_install_request': "10.77.132.127: Permission denied"
-    }
-    action_dict = {
-        'show_install_request': {'AFTER': 'disconnect'}
-    }
-
-    def disconnect(self):
-        self.RUNSHELL = False
-
-
-class TestNCS1KConnection(TestCase):
+class TestASR9KConnection(TestCase):
 
     def setUp(self):
-        self.server = TelnetServer(("127.0.0.1", 10023), NCS1KHandler)
+        self.server = TelnetServer(("127.0.0.1", 10023), ASR9KHandler)
         self.server_thread = Thread(target=self.server.serve_forever)
-        self.server_thread.setDaemon(True)
+        self.server_thread.daemon = True
         self.server_thread.start()
 
-        self.log_session = False
-        self.logfile_condoor = None  # sys.stderr
-        self.log_level = 0
+        debug = os.getenv("TEST_DEBUG", None)
+        if debug:
+            self.log_session = True
+            import sys
+            self.logfile_condoor = sys.stderr
+            self.log_level = 10
+
+        else:
+            self.log_session = False
+            self.logfile_condoor = None  # sys.stderr
+            self.log_level = 0
 
         try:
             os.remove('/tmp/condoor.shelve')
-        except:
+        except OSError:
             pass
 
     def tearDown(self):
-
+        self.conn.disconnect()
         self.server.shutdown()
         self.server.server_close()
-
         self.server_thread.join()
 
-    def test_NCS1K_1_discovery(self):
-
+    def test_ASR9K_1_discovery(self):
         urls = ["telnet://admin:admin@127.0.0.1:10023"]
         conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
         self.conn = conn
@@ -83,31 +74,36 @@ class TestNCS1KConnection(TestCase):
 
         self.assertEqual(conn._discovered, True, "Not discovered properly")
         self.assertEqual(conn.hostname, "ios", "Wrong Hostname: {}".format(conn.hostname))
-        self.assertEqual(conn.family, "NCS1K", "Wrong Family: {}".format(conn.family))
-        self.assertEqual(conn.platform, "NCS1002", "Wrong Platform: {}".format(conn.platform))
-        self.assertEqual(conn.os_type, "eXR", "Wrong OS Type: {}".format(conn.os_type))
-        self.assertEqual(conn.os_version, "6.0.1", "Wrong Version: {}".format(conn.os_version))
-        self.assertEqual(conn.udi['name'], "Rack 0", "Wrong Name: {}".format(conn.udi['name']))
-        self.assertEqual(conn.udi['description'], "Network Convergence System 1000 Controller",
+        self.assertEqual(conn.family, "ASR9K", "Wrong Family: {}".format(conn.family))
+        self.assertEqual(conn.platform, "ASR-9904", "Wrong Platform: {}".format(conn.platform))
+        self.assertEqual(conn.os_type, "XR", "Wrong OS Type: {}".format(conn.os_type))
+        self.assertEqual(conn.os_version, "5.3.3", "Wrong Version: {}".format(conn.os_version))
+        self.assertEqual(conn.udi['name'], "chassis ASR-9904-AC", "Wrong Name: {}".format(conn.udi['name']))
+        self.assertEqual(conn.udi['description'], "ASR 9904 2 Line Card Slot Chassis with V2 AC PEM",
                          "Wrong Description: {}".format(conn.udi['description']))
-        self.assertEqual(conn.udi['pid'], "NCS1002", "Wrong PID: {}".format(conn.udi['pid']))
+        self.assertEqual(conn.udi['pid'], "ASR-9904-AC", "Wrong PID: {}".format(conn.udi['pid']))
         self.assertEqual(conn.udi['vid'], "V01", "Wrong VID: {}".format(conn.udi['vid']))
-        self.assertEqual(conn.udi['sn'], "CHANGE-ME-", "Wrong S/N: {}".format(conn.udi['sn']))
+        self.assertEqual(conn.udi['sn'], "FOX1830GT5W", "Wrong S/N: {}".format(conn.udi['sn']))
         self.assertEqual(conn.prompt, "RP/0/RP0/CPU0:ios#", "Wrong Prompt: {}".format(conn.prompt))
-
         with self.assertRaises(condoor.CommandSyntaxError):
             conn.send("wrongcommand")
 
         conn.disconnect()
 
-    def test_NCS1K_2_connection_refused(self):
-        urls = ["telnet://admin:admin@127.0.0.1:10024"]
-        self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=0)
-        with self.assertRaises(condoor.ConnectionError):
-            self.conn.discovery(self.logfile_condoor)
-
-    def test_NCS1K_3_connection_wrong_user(self):
+    def test_ASR9K_2_connection_wrong_user(self):
         urls = ["telnet://root:admin@127.0.0.1:10023"]
-        self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=0)
+        self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
+
         with self.assertRaises(condoor.ConnectionError):
-            self.conn.discovery(self.logfile_condoor)
+            self.conn.connect(self.logfile_condoor)
+
+    def test_ASR9K_3_connection_refused(self):
+        urls = ["telnet://admin:admin@127.0.0.1:10024"]
+        self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
+        with self.assertRaises(condoor.ConnectionError):
+            self.conn.connect(self.logfile_condoor)
+
+
+if __name__ == '__main__':
+    from unittest import main
+    main()
