+++ tests/system/test_connection_ncs5500.py
@@ -28,32 +28,38 @@
 
 from unittest import TestCase
 
-from nxosmock.nxosmock import TelnetServer, NXOSHandler
+from tests.dmock.dmock import TelnetServer, NCS5500Handler
 from threading import Thread
 
 import condoor
 import os
 
 
-class NX9KHandler(NXOSHandler):
-    platform = "N9K"
-
-
-class TestNX9KConnection(TestCase):
-
+class TestNCS5500Connection(TestCase):
+    """
+    Test the device connection
+    """
     def setUp(self):
-        self.server = TelnetServer(("127.0.0.1", 10024), NX9KHandler)
+        self.server = TelnetServer(("127.0.0.1", 10023), NCS5500Handler)
         self.server_thread = Thread(target=self.server.serve_forever)
         self.server_thread.daemon = True
         self.server_thread.start()
 
-        self.log_session = False
-        self.logfile_condoor = None  # sys.stderr
-        self.log_level = 0
+        debug = os.getenv("TEST_DEBUG", None)
+        if debug:
+            self.log_session = True
+            import sys
+            self.logfile_condoor = sys.stderr
+            self.log_level = 10
+
+        else:
+            self.log_session = False
+            self.logfile_condoor = None  # sys.stderr
+            self.log_level = 0
 
         try:
             os.remove('/tmp/condoor.shelve')
-        except:
+        except OSError:
             pass
 
     def tearDown(self):
@@ -62,40 +68,39 @@ class TestNX9KConnection(TestCase):
         self.server.server_close()
         self.server_thread.join()
 
-    def test_NX9K_1_discovery(self):
-        urls = ["telnet://admin:admin@127.0.0.1:10024"]
+    def test_NCS5500_1_discovery(self):
+        urls = ["telnet://admin:admin@127.0.0.1:10023"]
         conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
         self.conn = conn
         conn.connect(self.logfile_condoor)
 
         self.assertEqual(conn._discovered, True, "Not discovered properly")
-        self.assertEqual(conn.hostname, "switch", "Wrong Hostname: {}".format(conn.hostname))
-        self.assertEqual(conn.family, "N9K", "Wrong Family: {}".format(conn.family))
-        self.assertEqual(conn.platform, "Nexus9000", "Wrong Platform: {}".format(conn.platform))
-        self.assertEqual(conn.os_type, "NX-OS", "Wrong OS Type: {}".format(conn.os_type))
-        self.assertEqual(conn.os_version, "7.0(3)IED5(1)", "Wrong Version: {}".format(conn.os_version))
-        self.assertEqual(conn.udi['name'], "Chassis", "Wrong Name: {}".format(conn.udi['name']))
-        self.assertEqual(conn.udi['description'], "Nexus9000 C9508 (8 Slot) Chassis",
+        self.assertEqual(conn.hostname, "ios", "Wrong Hostname: {}".format(conn.hostname))
+        self.assertEqual(conn.family, "NCS5500", "Wrong Family: {}".format(conn.family))
+        self.assertEqual(conn.platform, "NCS-5508", "Wrong Platform: {}".format(conn.platform))
+        self.assertEqual(conn.os_type, "eXR", "Wrong OS Type: {}".format(conn.os_type))
+        self.assertEqual(conn.os_version, "6.0.1", "Wrong Version: {}".format(conn.os_version))
+        self.assertEqual(conn.udi['name'], "Rack 0", "Wrong Name: {}".format(conn.udi['name']))
+        self.assertEqual(conn.udi['description'], "NCS5500 8 Slot Single Chassis",
                          "Wrong Description: {}".format(conn.udi['description']))
-        self.assertEqual(conn.udi['pid'], "N9K-C9508", "Wrong PID: {}".format(conn.udi['pid']))
+        self.assertEqual(conn.udi['pid'], "NCS-5508", "Wrong PID: {}".format(conn.udi['pid']))
         self.assertEqual(conn.udi['vid'], "V01", "Wrong VID: {}".format(conn.udi['vid']))
-        self.assertEqual(conn.udi['sn'], "FGE18210BQR", "Wrong S/N: {}".format(conn.udi['sn']))
-        self.assertEqual(conn.prompt, "switch#", "Wrong Prompt: {}".format(conn.prompt))
-        self.assertEqual(conn.is_console, True, "Console connection not detected")
+        self.assertEqual(conn.udi['sn'], "FGE194714QX", "Wrong S/N: {}".format(conn.udi['sn']))
+        self.assertEqual(conn.prompt, "RP/0/RP0/CPU0:ios#", "Wrong Prompt: {}".format(conn.prompt))
         with self.assertRaises(condoor.CommandSyntaxError):
             conn.send("wrongcommand")
 
         conn.disconnect()
 
-    def test_ASR9K_2_connection_wrong_user(self):
-        urls = ["telnet://root:admin@127.0.0.1:10024"]
+    def test_NCS5500_2_connection_wrong_user(self):
+        urls = ["telnet://root:admin@127.0.0.1:10023"]
         self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
 
         with self.assertRaises(condoor.ConnectionAuthenticationError):
             self.conn.connect(self.logfile_condoor)
 
-    def test_ASR9K_3_connection_refused(self):
-        urls = ["telnet://admin:admin@127.0.0.1:10023"]
+    def test_NCS5500_3_connection_refused(self):
+        urls = ["telnet://admin:admin@127.0.0.1:10024"]
         self.conn = condoor.Connection("host", urls, log_session=self.log_session, log_level=self.log_level)
         with self.assertRaises(condoor.ConnectionError):
             self.conn.connect(self.logfile_condoor)
