+++ tox.ini
@@ -18,7 +18,7 @@ changedir = tests
 
 commands =
     coverage run -m unittest discover -s . -p test*.py -t .. -v []
-    coverage report --fail-under=40 --omit=test*,*.tox*.py,*mock*
+    coverage report --fail-under=40 --omit=test*,*.tox*.py,*mock*,system/*,unit/*
 
 [testenv:py27]
